fix: move textentry to separate component
