+++ pages/addpage.tsx
@@ -1,49 +1,13 @@
 import "../app/globals.css";
 import Header from "@/components/header";
 import Image from "next/image";
+import TextEntry from "@/components/textentry";
 import pfp  from '../public/pfp.png';
 import { FormEvent, useState } from "react";
 
-interface TextEntryProps {
-    label: string;
-    id: string;
-    placeholder?: string;
-    value: string;
-    changeHandler: (event: FormEvent) => void;
-}
-
-function TextEntry(props: TextEntryProps){
-    return (
-        <div>
-            <label htmlFor={props.id} className="block text-sm font-medium leading-6 text-gray-900">
-                {props.label}
-            </label>
-            <div className="flex rounded-md shadow-sm ring-1 ring-inset ring-gray-300 focus-within:ring-2 focus-within:ring-inset focus-within:ring-indigo-600 sm:max-w-md mt-2">
-                <input
-                    type="text"
-                    name={props.id}
-                    id={props.id}
-                    value={props.value}
-                    onChange={props.changeHandler}
-                    autoComplete={props.id}
-                    className="block flex-1 border-0 bg-transparent py-1.5 pl-1 text-gray-900 placeholder:text-gray-400 focus:ring-0 sm:text-sm sm:leading-6"
-                    placeholder={props.placeholder}
-                />
-            </div>
-        </div>
-    );
-};
-
-function handleSubmit(){
-    /* handle form submission */
-}
-
-function handleCancel(){
-    /* handle form cancellation */
-}
-
 export default function AddPage(){
     const [contact, setContact] = useState({
+        photo: "",
         college: "",
         major: "",
         firstName: "",
@@ -66,9 +30,37 @@ export default function AddPage(){
         }
     });
 
+    function handleSubmit(){
+        /* handle form submission */
+        // postContact()
+        // .then((res : Response) => {
+        //     if(res.status === 201){
+        //         return res.json();
+        //     }
+        //     else{
+        //         return undefined;
+        //     }
+        // })
+        // .catch((error : Error) => {
+        //     console.log(error);
+        // });
+        console.log(contact);
+        console.log("test");
+    }
+    
+    function handleCancel(){
+        /* handle form cancellation */
+        // redirect to user home page
+    }
+
     function handleChange(event: FormEvent){
         const {name, value} = event.target as HTMLFormElement;
         switch(name){
+            case "file-upload": {
+                setContact({...contact, photo: URL.createObjectURL(value.files[0])});
+                console.log(contact.photo);
+                break;
+            }
             case "college": {
                 setContact({...contact, college: value});
                 break;
@@ -150,12 +142,12 @@ export default function AddPage(){
 
                             <div className="mt-10 grid grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6">
                                 <div className="col-span-full">
-                                    <label htmlFor="photo" className="block text-sm font-medium leading-6 text-gray-900">
+                                    <div className="block text-sm font-medium leading-6 text-gray-900">
                                         Photo
-                                    </label>
+                                    </div>
                                     <div className="mt-2 flex items-center gap-x-3">
                                         <Image 
-                                            src={pfp}
+                                            src={contact.photo ? contact.photo : pfp}
                                             width= "50"
                                             height= "50"
                                             alt="upload profile picture preview"
@@ -360,12 +352,12 @@ export default function AddPage(){
                                 <button type="button" className="text-sm font-semibold leading-6 text-gray-900">
                                     Cancel
                                 </button>   
-                                <button
-                                    type="submit"
+                                <input 
+                                    type = "submit"
+                                    value="Submit"
                                     className="rounded-md bg-indigo-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
-                                >
-                                    Save
-                                </button>
+                                    onClick={handleSubmit}
+                                />
                             </div>
                         </div>
                     </div>
